Code cleanup - Move TNC classes( ArrivalTime, RideEstimate, TransportationNetworkCompany, TransportationNetworkCompanyService) from main to Sandbox.
