+++ src/ext-test/java/org/opentripplanner/ext/tnc/updater/LyftTransportationNetworkCompanyDataSourceTest.java
@@ -5,10 +5,10 @@ import com.github.tomakehurst.wiremock.junit.WireMockRule;
 import org.junit.Before;
 import org.junit.Rule;
 import org.junit.Test;
+import org.opentripplanner.ext.tnc.routing.model.ArrivalTime;
+import org.opentripplanner.ext.tnc.routing.model.RideEstimate;
 import org.opentripplanner.ext.tnc.updater.lyft.LyftAuthenticationRequestBody;
 import org.opentripplanner.ext.tnc.updater.lyft.LyftTransportationNetworkCompanyDataSource;
-import org.opentripplanner.routing.transportation_network_company.ArrivalTime;
-import org.opentripplanner.routing.transportation_network_company.RideEstimate;
 
 import java.io.IOException;
 import java.util.List;
