+++ src/ext-test/java/org/opentripplanner/ext/tnc/updater/NoApiTransportationNetworkCompanyDataSourceTest.java
@@ -1,8 +1,8 @@
 package org.opentripplanner.ext.tnc.updater;
 
 import org.junit.Test;
-import org.opentripplanner.routing.transportation_network_company.ArrivalTime;
-import org.opentripplanner.routing.transportation_network_company.RideEstimate;
+import org.opentripplanner.ext.tnc.routing.model.ArrivalTime;
+import org.opentripplanner.ext.tnc.routing.model.RideEstimate;
 
 import java.io.IOException;
 import java.util.List;
