+++ src/ext-test/java/org/opentripplanner/ext/tnc/updater/UberTransportationNetworkCompanyDataSourceTest.java
@@ -3,9 +3,9 @@ package org.opentripplanner.ext.tnc.updater;
 import com.github.tomakehurst.wiremock.junit.WireMockRule;
 import org.junit.Rule;
 import org.junit.Test;
+import org.opentripplanner.ext.tnc.routing.model.ArrivalTime;
+import org.opentripplanner.ext.tnc.routing.model.RideEstimate;
 import org.opentripplanner.ext.tnc.updater.uber.UberTransportationNetworkCompanyDataSource;
-import org.opentripplanner.routing.transportation_network_company.ArrivalTime;
-import org.opentripplanner.routing.transportation_network_company.RideEstimate;
 
 import java.io.IOException;
 import java.util.List;
