+++ src/ext/java/org/opentripplanner/ext/tnc/routing/model/RideEstimate.java
@@ -1,4 +1,4 @@
-package org.opentripplanner.routing.transportation_network_company;
+package org.opentripplanner.ext.tnc.routing.model;
 
 /**
  * A class to model the estimated ride time while using service from a transportation network
