+++ src/ext/java/org/opentripplanner/ext/tnc/updater/uber/UberTransportationNetworkCompanyDataSource.java
@@ -3,12 +3,12 @@ package org.opentripplanner.ext.tnc.updater.uber;
 import com.fasterxml.jackson.databind.DeserializationFeature;
 import com.fasterxml.jackson.databind.JsonNode;
 import com.fasterxml.jackson.databind.ObjectMapper;
+import org.opentripplanner.ext.tnc.routing.model.ArrivalTime;
+import org.opentripplanner.ext.tnc.routing.model.RideEstimate;
+import org.opentripplanner.ext.tnc.routing.model.TransportationNetworkCompany;
 import org.opentripplanner.ext.tnc.updater.Position;
 import org.opentripplanner.ext.tnc.updater.RideEstimateRequest;
 import org.opentripplanner.ext.tnc.updater.TransportationNetworkCompanyDataSource;
-import org.opentripplanner.routing.transportation_network_company.ArrivalTime;
-import org.opentripplanner.routing.transportation_network_company.RideEstimate;
-import org.opentripplanner.routing.transportation_network_company.TransportationNetworkCompany;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
